[Create][Fix] adjust service folder and build up an axios workflow
