+++ project/frontend/.env.example
@@ -1,0 +1,1 @@
+API_BASE_URL = 'your backend base URL'
