+++ project/frontend/.gitignore
@@ -28,3 +28,6 @@ coverage
 *.sw?
 
 *.tsbuildinfo
+
+# custom
+.env
