+++ project/frontend/src/service/LibraryApi/QuizAPI.js
@@ -1,0 +1,3 @@
+import api from '@/service/ApiClient';
+
+// see FetchQuestion() and Created() in UserLibrary.vue to know waht api to call
